Split composer Playwright flows
